delete four files to reduce update time
