--- app/constants.py
@@ -1,4 +1,0 @@
-longpress_ms = 1500
-
-socket_connect_s = 2
-request_timeout_s = 5
