--- app/run.py
@@ -1,1 +1,0 @@
-from . import start as _
